+++ src/control/Button.tsx
@@ -4,11 +4,14 @@ import { Box, IBoxProps } from "../layout/content/box/Box";
 import { IBoxColorsThunk } from "../layout/content/box/IBoxColorsThunk";
 import { HoverState } from "../util/react/HoverState";
 import { ITheme } from "../theme/ITheme";
+import { IBoxColors } from "../layout/content/box/IBoxColors";
+import Color from "color";
 
 interface IButtonRootProps {
-    floating?: boolean;
+    elevation?: Elevation;
     disabled?: boolean;
     rounded?: boolean;
+    shadowColor(theme: ITheme): string | undefined;
 }
 
 const ButtonRoot = styled<IButtonRootProps, "button">("button")`
@@ -17,11 +20,13 @@ const ButtonRoot = styled<IButtonRootProps, "button">("button")`
     ` : ``}
     user-select: none;
     text-transform: uppercase;
-    ${props => props.floating ? css`
-    box-shadow: 0 8px 16px rgba(167, 181, 209, 0.6);
-    ` : ``}
 
-    border-radius: ${props => props.rounded ? "100px" : "4px" };
+    ${props => props.elevation !== "none" ? css`
+    box-shadow: ${props.elevation === "high" ? "0 8px 16px" : "0 4px 8px"}
+        ${Color(props.shadowColor(props.theme) || "transparent").alpha(0.6).toString()}
+    ` : ``};
+
+    border-radius: ${props => props.rounded ? "100px" : "4px"};
 
     /* Override user-agent styles */
     background: none;
@@ -35,11 +40,39 @@ const StyledBox = styled(Box)`
     transition: background-color .2s ease-in-out, border-color .2s ease-in-out;
 `;
 
-export type ButtonColors = "primary" | "secondary" | "special1" | "special2";
+export type ButtonColors = "primary" | "secondary" | "special";
+type Elevation = "none" | "high" | "low";
 type InteractionState = "normal" | "hover" | "disabled";
-type GetColorSetFn = (colorVariants: ButtonColors, state: InteractionState) => IBoxColorsThunk<ITheme>;
+type GetColorSetFn = (colorVariants: ButtonColors, state: InteractionState)
+    => IBoxColorsThunk<ITheme>;
 
 const getColorSet: GetColorSetFn = (colorVariant, state) => (theme) => theme.colors.button[colorVariant][state];
+
+const getColors = (colors: IBoxColors | IBoxColorsThunk<ITheme>, theme: ITheme) => {
+    if (typeof colors === "function") {
+        return colors(theme);
+    }
+    return colors;
+};
+
+const getState = (disabled: boolean, hover: boolean) => {
+    if (disabled) {
+        return "disabled";
+    } else if (hover) {
+        return "hover";
+    } else {
+        return "normal";
+    }
+};
+
+const invertColors: GetColorSetFn = (colorVariant, state) => (theme) => {
+    const colors = theme.colors.button[colorVariant][state];
+    return {
+        text: colors.background!,
+        background: colors.text,
+        border: colors.background
+    };
+};
 
 type IHtmlButtonProps = Pick<React.ButtonHTMLAttributes<HTMLButtonElement>, "autoFocus" | "name" | "type" | "value">;
 
@@ -47,52 +80,65 @@ export interface IButtonProps extends IHtmlButtonProps {
     Icon?: IBoxProps["Icon"];
     iconPlacement?: IBoxProps["iconPlacement"];
     colors?: ButtonColors;
-    floating?: boolean;
+    /** Only for primary button */
+    elevation?: Elevation;
     /** Rounded 50% corners */
     rounded?: boolean;
+    /** Only for primary button */
+    inverted?: boolean;
     children?: string;
     disabled?: boolean;
     onClick?(): void;
 }
 
 export class Button extends React.Component<IButtonProps> {
-    static defaultProps: Pick<IButtonProps, "colors"> = {
-        colors: "secondary"
+    static defaultProps: Pick<IButtonProps, "colors" | "elevation"> = {
+        colors: "secondary",
+        elevation: "none"
     };
 
     render() {
-        let {
-            Icon, iconPlacement, floating, disabled, rounded, colors, autoFocus, name, type, value, children
-        } = this.props;
+        let { Icon, iconPlacement, elevation, disabled, rounded,
+            inverted, colors, autoFocus, name, type, value, children } = this.props;
+
         return (
             <HoverState>
-                {(hover) =>
-                <ButtonRoot
-                    onClick={!this.props.disabled ? this.props.onClick : void 0}
-                    floating={floating}
-                    disabled={disabled}
-                    rounded={rounded}
-                    autoFocus={autoFocus}
-                    name={name}
-                    type={type}
-                    value={value}
-                >
-                    <StyledBox
-                        Icon={Icon}
-                        iconPlacement={iconPlacement ? iconPlacement : "left"}
-                        colors={getColorSet(colors!, !disabled ? hover ? "hover" : "normal" : "disabled")}
-                        metrics={{
-                            fontSize: 12,
-                            lineHeight: 14,
-                            fontWeight: 600,
-                            height: 32,
-                            iconSize: 24,
-                            letterSpacing: .4,
-                            textPaddingTop: 7,
-                            textPaddingX: 16
-                        }}
-                    >{children}</StyledBox>
-                </ButtonRoot>
+                {(hover) => {
+                    let state: InteractionState = getState(disabled!, hover);
+                    let colorSet = getColorSet(colors!, state);
+                    let isInverted = state === "normal" && colors === "primary" && inverted;
+                    let invertedColorSet = invertColors(colors!, state);
+                    let shadowColor = (theme: ITheme) =>
+                        colors === "primary" ? getColors(colorSet, theme).background : void 0;
+
+                    return <ButtonRoot
+                        onClick={!this.props.disabled ? this.props.onClick : void 0}
+                        elevation={elevation}
+                        shadowColor={shadowColor}
+                        disabled={disabled}
+                        rounded={rounded}
+                        autoFocus={autoFocus}
+                        name={name}
+                        type={type}
+                        value={value}
+                    >
+                        <StyledBox
+                            Icon={Icon}
+                            iconPlacement={iconPlacement ? iconPlacement : "left"}
+                            colors={isInverted ? invertedColorSet : colorSet}
+                            metrics={{
+                                fontSize: 12,
+                                lineHeight: 14,
+                                fontWeight: 600,
+                                height: 32,
+                                iconSize: 24,
+                                letterSpacing: .4,
+                                textPaddingTop: 7,
+                                textPaddingX: 16
+                            }}
+                        >{children}</StyledBox>
+                    </ButtonRoot>;
+                }
                 }
             </HoverState>
         );